% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{adj_to_tpo}
\alias{adj_to_tpo}
\title{Convert an adjacency matrix to a topology}
\usage{
adj_to_tpo(adj)
}
\arguments{
\item{adj}{the adjacency martix to be converted to a topology file}
}
\description{
Convert an adjacency matrix to a topology
}
