% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all.circuits}
\alias{all.circuits}
\title{Data file containing the topology of all 60212 non-redundant four-node gene circuits}
\format{
A list of 60212 circuits, each element is a matrix of adjacency matrix. 
Four columns/rows represent genes "A", "B", "C", and "D".
In each adjacency matrix, 1 represents activation, 2 represents inhibition, and 0 represents no interaction.
}
\usage{
all.circuits
}
\description{
This data contains the topology of all 60212 non-redundant four-node gene circuits.
}
\keyword{datasets}
