% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis_circuit_2node}
\alias{analysis_circuit_2node}
\title{Analysis script to evaluate the state distribution of a two-node circuit motif}
\usage{
analysis_circuit_2node(tpo, numModels = 10000)
}
\arguments{
\item{tpo}{Data frame. Topology data of a two-node circuit motif. 
The data frame contains three columns: Source, Target, Interaction type}

\item{numModels}{Numeric. Number of models to be simulated. Default: 10000}
}
\description{
Analysis script to evaluate the state distribution of a two-node circuit motif
}
