% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{circuit_grouping}
\alias{circuit_grouping}
\title{Grouping two-node circuit motifs by their types}
\usage{
circuit_grouping()
}
\value{
Vector of factor: types for all two-node motifs.
}
\description{
Grouping two-node circuit motifs by their types
}
