% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{dist_ks}
\alias{dist_ks}
\title{Calculate th KS distance of two gene expression distributions}
\usage{
dist_ks(query, reference, experimental)
}
\arguments{
\item{query}{sRACIPE object or PCA matrix of the query data}

\item{reference}{sRACIPE object of the reference data}

\item{experimental}{Logical. T: query is the PCA matrix from an experimental dataset. F: query is an sRACIPE object}
}
\value{
the distance between the gene expression distributions from the query and reference.
}
\description{
Calculate th KS distance of two gene expression distributions
}
