% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{enrichment_coupling}
\alias{enrichment_coupling}
\title{Motif enrichment analysis for the coupling of two two-node circuit motifs}
\usage{
enrichment_coupling(
  all.circuits = all.circuits,
  all.scores,
  motif_list,
  new_ind,
  overlap_list,
  decreasing = T,
  if_overlap = 2,
  topCircuits = 600,
  nhill = 20
)
}
\arguments{
\item{all.circuits}{List of the topologies of all 60212 non-redundant 4-node circuits. Default "all.circuits" from the package data.}

\item{all.scores}{Data frame. 1st column: scores; 2nd column: circuit index. The data frame is ordered by the scores.}

\item{motif_list}{List of 2 by 2 integer matrix. 2-node motif info generated by the function: generate_motif_list.}

\item{new_ind}{Vector of integer. mapping to non-redundant 2-node motifs generated by the function: generate_index_conversion.}

\item{overlap_list}{Matrix (6 by 6 integers). Info of overlapping from the motif location in the 4-node circuit
generated by the function: generate_overlap_data.}

\item{decreasing}{Logic. Whether circuits are ranked by the scores in a decreasing order (T) or not (F). Default T.}

\item{if_overlap}{Whether consider two motifs with overlapping (1), without overlapping (0), or both (2). Default 2.}

\item{topCircuits}{Integer. Number of top circuits for the enrichment analysis. Default 600.}

\item{nhill}{Integer/Numeric. Hill coefficient for the Hill function as the weighting factor for motif counts, 
large n makes the Hill function more binary. Default 20.}
}
\value{
A data frame containing the enrichment scores of the coupling between 2-node circuit motifs (39 by 39)
}
\description{
Motif enrichment analysis for the coupling of two two-node circuit motifs
}
