% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{gen_network_scalefree}
\alias{gen_network_scalefree}
\title{Generate a random scale-free gene network consisting the two-node motifs of choice}
\usage{
gen_network_scalefree(num_nodes, motif_list, motif_choice)
}
\arguments{
\item{num_nodes}{Integer. Number of nodes of the gene network to be generated.}

\item{motif_list}{List of 2 by 2 integer matrix. 2-node motif info generated by the function: generate_motif_list.}

\item{motif_choice}{Vector of integer. A vector of indices of the selected two-node circuit motifs}
}
\value{
Matrix. Adjacency matrix of the generated network
}
\description{
Generate a random scale-free gene network consisting the two-node motifs of choice
}
