% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{generate_index_conversion}
\alias{generate_index_conversion}
\title{Convert redundant indices to the indices for all 39 non-redundant motifs}
\usage{
generate_index_conversion()
}
\description{
Convert redundant indices to the indices for all 39 non-redundant motifs
}
