% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{generate_motif_list}
\alias{generate_motif_list}
\title{Generate a list of all 72 2-node circuit motifs (including redundant ones)}
\usage{
generate_motif_list()
}
\value{
list of all 72 2-node circuit motifs
}
\description{
Generate a list of all 72 2-node circuit motifs (including redundant ones)
}
