% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{get_motif_adj}
\alias{get_motif_adj}
\title{Obtain the adjacency matrix for a specific two-node circuit motif (number should be from 1 to 39)}
\usage{
get_motif_adj(number, motif_list)
}
\arguments{
\item{number}{Integer. Index of the two-node circuit motif.}

\item{motif_list}{List of 2 by 2 integer matrix. 2-node motif info generated by the function: generate_motif_list.}
}
\value{
Integer matrix. 2 by 2 adjacency matrix.
}
\description{
Obtain the adjacency matrix for a specific two-node circuit motif (number should be from 1 to 39)
}
