% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{lin_score}
\alias{lin_score}
\title{The scoring function for ranking circuits with a linear state distribution}
\usage{
lin_score(rset)
}
\arguments{
\item{rset}{The sRACIPE object of the simulated circuit.}
}
\value{
return(min(score_vector)): returns the score for the linear state distribution.
}
\description{
The scoring function for ranking circuits with a linear state distribution
}
