% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{map_and_project}
\alias{map_and_project}
\title{Compare the state distributions of two four-node circuits, find the most matched genes, 
and project the simulated gene expression data of the 2nd circuit to the PCs of the 1st circuit}
\usage{
map_and_project(rset1, rset2)
}
\arguments{
\item{rset1}{sRACIPE object. RACIPE simulation data for the first circuit}

\item{rset2}{sRACIPE object. RACIPE simulation data for the second circuit}
}
\value{
Numeric matrix. PC coordinates of RACIPE simulated gene expression of 
the second circuit projected onto the PCs of the RACIPE simulated gene expression of 
the first circuit
}
\description{
Compare the state distributions of two four-node circuits, find the most matched genes, 
and project the simulated gene expression data of the 2nd circuit to the PCs of the 1st circuit
}
