% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif_analysis}
\alias{motif_analysis}
\title{A combined motif enrichment analysis for single two-node circuit motifs and motif coupling}
\usage{
motif_analysis(
  all.circuits = all.circuits,
  all.scores,
  ylim = NULL,
  color_breaks = NULL,
  filename = NULL,
  decreasing = T,
  topCircuits = 600
)
}
\arguments{
\item{all.circuits}{List of the topologies of all 60212 non-redundant 4-node circuits. Default "all.circuits" from the package data.}

\item{all.scores}{Data frame. 1st column: scores; 2nd column: circuit index. The data frame is ordered by the scores.}

\item{ylim}{Vector of numerics (2). Y axis limit for single motif enrichment. Default: NULL.}

\item{color_breaks}{Vector that defines color scaling for pheatmap. Default: NULL}

\item{filename}{Character. Prefix of filenames for plotting. Default: NULL. If provided, plots are also saved to files.}

\item{decreasing}{Logical. Whether circuits are ranked by the scores in a decreasing order (T) or not (F). Default T.}

\item{topCircuits}{Integer. Number of top circuits for the enrichment analysis. Default 600.}
}
\value{
List of plotting objects for single motif and motif coupling enrichment analyses.
}
\description{
A combined motif enrichment analysis for single two-node circuit motifs and motif coupling
}
