% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_adj}
\alias{plot_adj}
\title{Plot a network from the adjacency matrix}
\usage{
plot_adj(adj)
}
\arguments{
\item{adj}{Matrix (4 by 4, integer). Adjacency matrix of a four node circuit.}
}
\value{
empty
}
\description{
Plot a network from the adjacency matrix
}
