% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_motif}
\alias{plot_motif}
\title{Plot a specific circuit motif}
\usage{
plot_motif(number, motif_list)
}
\arguments{
\item{number}{Integer. Index of a two-node circuit motif}

\item{motif_list}{List of 2 by 2 integer matrix. 2-node motif info generated by the function: generate_motif_list.}
}
\value{
empty
}
\description{
Plot a specific circuit motif
}
