% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_net}
\alias{plot_net}
\title{Network plotting function}
\usage{
plot_net(tf_links = tf_links)
}
\arguments{
\item{tf_links}{Data frame of circuit edge list. 
Three columns: Source, Target, Interaction Types -- 1: Activation; 2: Inhibition}
}
\value{
empty
}
\description{
Network plotting function
}
