% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_RACIPE}
\alias{plot_RACIPE}
\title{Scatterplots of PCA and Gene expression from any 4-node circuit}
\usage{
plot_RACIPE(rset)
}
\arguments{
\item{rset}{The sRACIPE object of the simulated circuit}
}
\value{
list(g,p)
}
\description{
Scatterplots of PCA and Gene expression from any 4-node circuit
}
