% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{sim_4node}
\alias{sim_4node}
\title{Simulate one of the 4-node circuits from the 60212 unique circuits}
\usage{
sim_4node(index, Gaussian = F, numModels = 10000, all.circuits = all.circuits)
}
\arguments{
\item{index}{Numeric index number of 4-node circuit to be simulate. Takes values from 1:60212}

\item{Gaussian}{Logical. If T, kinetic parameters will be sampled from a gaussian distribution. If F, kinetic parameters will be sampled from a uniform distribution}

\item{numModels}{Numeric. Number of models to be simulated. Default: 10000}

\item{all.circuits}{List. The topology of all 60212 circuit motifs. Default "all.circuits" from the package data.}
}
\value{
rset: sRACIPE object. RACIPE simulation results for a circuit
}
\description{
Simulate one of the 4-node circuits from the 60212 unique circuits
}
