% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{simu_rnorm}
\alias{simu_rnorm}
\title{RACIPE simulations with random kinetic parameters from Gaussian distributions}
\usage{
simu_rnorm(rset, numModels = 10000)
}
\arguments{
\item{rset}{sRACIPE object. sRACIPE output from a standard RACIPE simulation}

\item{numModels}{Numeric. Number of models to be simulated. Default: 10000.}
}
\description{
RACIPE simulations with random kinetic parameters from Gaussian distributions
}
