% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{single_motif_permute}
\alias{single_motif_permute}
\title{Generate permutations for p-value calculations}
\usage{
single_motif_permute(
  all.circuits = all.circuits,
  all.scores,
  decreasing = T,
  topCircuits = 600,
  nhill = 20,
  num_perm = 1000
)
}
\arguments{
\item{all.circuits}{List of the topologies of all 60212 non-redundant 4-node circuits. Default "all.circuits" from the package data.}

\item{all.scores}{Data frame. 1st column: scores; 2nd column: circuit index. The data frame is ordered by the scores.}

\item{decreasing}{Logical. Whether circuits are ranked by the scores in a decreasing order (T) or not (F). Default T.}

\item{topCircuits}{Integer. Number of top circuits for the enrichment analysis. Default 600.}

\item{nhill}{Integer/Numeric. Hill coefficient for the Hill function as the weighting factor for motif counts, 
large n makes the Hill function more binary. Default 20.}

\item{num_perm}{Integer. Number of permutations. (1000)}
}
\value{
Numeric vector. Adjusted p values for the enrichement of each two-node circuit motif (a total of 39)
}
\description{
Generate permutations for p-value calculations
}
