% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{trig_score}
\alias{trig_score}
\title{The scoring function for ranking circuits with a triangular state distribution}
\usage{
trig_score(rset)
}
\arguments{
\item{rset}{The sRACIPE object of the simulated circuit.}
}
\value{
return(min(score_vector)): returns the score for the triangular state distribution.
}
\description{
The scoring function for ranking circuits with a triangular state distribution
}
