% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{z_score}
\alias{z_score}
\title{Calculate z-score for a specific distance}
\usage{
z_score(distances, score)
}
\arguments{
\item{distances}{list of distances}

\item{score}{Numeric. score to calculate the z-score}
}
\value{
z-score
}
\description{
Calculate z-score for a specific distance
}
