% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis_circuit_2node}
\alias{analysis_circuit_2node}
\title{Analysis script to evaluate the state distribution of a two-node gene circuit}
\usage{
analysis_circuit_2node(rSet, numModels = 10000, filename)
}
\arguments{
\item{rSet}{sRACIPE object. RACIPE simulation data.}

\item{numModels}{Numeric. Number of models to be simulated. Default: 10000}

\item{filename}{Character. Filename. Default: "myplot".}
}
\description{
Analysis script to evaluate the state distribution of a two-node gene circuit
}
